Clean or Working Project
I find myself wanting to use this framework for a current project so that I can see it in action and find out the needs, but I fear that it will become cluttered. How can I setup Liriope so that it can be updated and not break the site it is being used on? If the framework files are contained in a folder, this may work, but the implied directory structure must be in place. Ug... we'll see.