Minor tweaks
I wanted to note a TODO about rewrite. While at it, I tweaked the contact us controller which should go away sometime. Oh, and the reference to the LESS CSS js file was not from the root, so I fixed that when checking the URL with index.php thrown in.