Controller > View > page > theme
I was having a hard time tracking the logic path so I've now compartmentalized code into classes and built them into the controller and view process. Now, the controller feeds it's view file to a page object and the view model then renders that page inside of a theme object which defaults to an included theme. Ug... this has to be more simple that I'm making it.