Found the injection point
Took some digging. I have written something compliated! But I found the point where the page content is collected into a variable, which is exactly where I'd start the filtering. So, search for the TODO and that's where it is. Oh, and while I was digging, I changed theme PHP files from theme name to index.