We have filters
Granted, the two that I've put in place are silly at best, but I've proven that it's possible. More work is needed for sorting, and the slugs aren't used... hmm... I need to use those for the filter's array. Anywho... they're in! The days of inline CSS and comment removal are close at hand.
The Liriope set of filters need to be housed elsewhere. Perhaps in the filter.class.php file rather than the seed.php file. But for now, they are in and they work.