HTACCESS... beefy or lean?
I started this project by following tutorials. In one, the htaccess file was put together. I've since cut it back severely, but not totally. I'm still unsure how much is needed, or whether it is in the scope of this framework to require one.

Truly, the only addition to the .htaccess file that I do need is the rewriting. Perhaps I should cut it further to only this.