No more intro text
Now the blog full article page can turn off the intro text using a configuration setting. Defaults to TRUE. This way, the blog article is simply the content. Ah.

Oh, and the blog title linking to the full article, it only affects the H1 tags, so this could be why it's not working on all blog articles.