Testing router
Well, I learned a bit more about how what I've done works. Sad, but true. The router needs to have abosolute paths from the web root to work properly. And all images need to be referenced as if index.php is throwing them (cuz it is)! So that means src="/content/..." needs to be at the front of images in your pages.