Working as is, but so many TODOs added!
I have roughed in changing View to static and Page to public, but now I wonder if I need to extend my models with Page!
I'd love to be able to set "page" variables within the content file. This file is "rendered" withing the page object
so in theory, you can set self::$variable or $this->variable, but then, with other models, the content is rather
read with those model objects, and the content is dumped to Page, so then the context is wrong! Hmm... this is a
null issue for single pages using Folderfile, but Blogs, and Blogposts?! Ug.